Fix some integer type bugs

+++ problems/src/problem_003.rs
@@ -1,7 +1,7 @@
 use utils;
 
-pub fn problem_003() -> usize {
-    let n: usize = 600851475143;
+pub fn problem_003() -> u64 {
+    let n: u64 = 600851475143;
     *utils::prime_factors(n).iter().max().unwrap()
 }
 
@@ -12,7 +12,7 @@ mod test {
 
     #[test]
     fn test_problem_003() {
-        let ans: usize = problem_003();
+        let ans: u64 = problem_003();
         println!("Answer to Problem 3: {}", ans);
         assert!(ans == 6857)
     }
